Prepare for 2.0.0 release

+++ jackson/src/main/scala/com/twitter/finatra/json/internal/caseclass/validation/validators/SizeValidator.scala
@@ -55,7 +55,7 @@ class SizeValidator(
     }
 
     ValidationResult(
-      isValid(size),
+      isValid(size.toLong),
       errorMessage(
         validationMessageResolver,
         value,
